+++ src/app/components/loader/loader.component.ts
@@ -31,7 +31,7 @@
 */
 
 // Angular imports
-import { Component, OnInit, Input, OnChanges } from "@angular/core";
+import { Component, input } from "@angular/core";
 
 @Component({
   selector: "app-loader",
@@ -39,100 +39,10 @@ import { Component, OnInit, Input, OnChanges } from "@angular/core";
   styleUrl: "./loader.component.less",
   standalone: true,
 })
-export class Loader implements OnInit, OnChanges {
-  // the sort of data the parent component is waiting for
-  @Input() waitingFor?: string;
-  // message to display to the user while waiting
-  message = "";
+export class Loader {
+  // the message to display to the user while waiting
+  loadingMessage = input<string>("Loading...");
 
   // CTOR
   constructor() {}
-
-  /*
-  Function Name: ngOnInit()
-  Function Description: This method is automatically triggered by Angular upon
-                        page initiation. It checks for the existence of the data
-                        the parent component is waiting (via checkLoadingTarget()) for and adjusts the
-                        loader itself (its message and visibility) accordingly.
-  Parameters: None.
-  ----------------
-  Programmer: Shir Bar Lev.
-  */
-  ngOnInit() {
-    this.checkLoadingTarget();
-  }
-
-  /*
-  Function Name: ngOnChanges()
-  Function Description: This method is automatically triggered by Angular upon
-                        changes in parent component. It checks for the existence of the data
-                        the parent component is waiting (via checkLoadingTarget()) for and adjusts the
-                        loader itself (its message and visibility) accordingly.
-  Parameters: None.
-  ----------------
-  Programmer: Shir Bar Lev.
-  */
-  ngOnChanges() {
-    if (this.waitingFor) {
-      this.checkLoadingTarget();
-    }
-  }
-
-  /*
-  Function Name: checkLoadingTarget()
-  Function Description: Checks for the existence of the data the parent component
-                        is waiting (via checkLoadingTarget()) for and adjusts the
-                        loader itself (its message and visibility) accordingly.
-  Parameters: None.
-  ----------------
-  Programmer: Shir Bar Lev.
-  */
-  checkLoadingTarget() {
-    // if the app is waiting for user data to be fetched from the server
-    if (this.waitingFor == "user" || this.waitingFor == "other user") {
-      this.message = "Fetching user data...";
-    }
-    // if the app is waiting for incoming/outgoing messages data to be
-    // fetched from the server
-    else if (
-      this.waitingFor == "inbox messages" ||
-      this.waitingFor == "outbox messages" ||
-      this.waitingFor == "thread messages"
-    ) {
-      this.message = "Fetching messages...";
-    }
-    // if the app is waiting for threads messages data to be fetched from the server
-    else if (this.waitingFor == "threads messages") {
-      this.message = "Fetching threads...";
-    }
-    // if the app is waiting for main page/new posts/suggested
-    // posts data to be fetched from the server
-    else if (
-      this.waitingFor == "main page" ||
-      this.waitingFor == "new posts" ||
-      this.waitingFor == "suggested posts"
-    ) {
-      this.message = "Fetching posts...";
-    }
-    // if the app is waiting for the user's posts to be fetched from the server
-    else if (this.waitingFor == "user posts") {
-      this.message = "Fetching user posts...";
-    }
-    // if the app is waiting for search results to be fetched from the server
-    else if (this.waitingFor == "search") {
-      this.message = "Searching...";
-    }
-    // if the app is waiting for open reports to be fetched from the server
-    else if (this.waitingFor == "admin reports") {
-      this.message = "Getting user and post reports...";
-    }
-    // if the app is waiting for blocked users to be fetched from the server
-    else if (this.waitingFor == "admin blocks") {
-      this.message = "Getting blocked users...";
-    }
-    // if the app is waiting for filtered phrases to be fetched from the server
-    else if (this.waitingFor == "admin filters") {
-      this.message = "Getting filtered phrases...";
-    }
-  }
 }